+++ apps/web/dynastyweb/src/app/admin/login/page.tsx
@@ -1,164 +1,10 @@
-'use client';
-
-import { useState } from 'react';
-import { useRouter, useSearchParams } from 'next/navigation';
-import Link from 'next/link';
-import { signInWithEmailAndPassword } from 'firebase/auth';
-import { auth } from '@/lib/firebase';
-import { adminSecurity } from '@/utils/adminSecurity';
-import { Shield, Loader2, AlertCircle } from 'lucide-react';
-import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import { Alert, AlertDescription } from '@/components/ui/alert';
-import { useToast } from '@/components/ui/use-toast';
+import { Suspense } from 'react';
+import AdminLoginClient from './AdminLoginClient';
 
 export default function AdminLoginPage() {
-  const router = useRouter();
-  const searchParams = useSearchParams();
-  const { toast } = useToast();
-  
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [isLoading, setIsLoading] = useState(false);
-  const [error, setError] = useState('');
-
-  const redirect = searchParams.get('redirect') || '/dashboard';
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setIsLoading(true);
-    setError('');
-
-    try {
-      // Sign in with Firebase
-      await signInWithEmailAndPassword(auth, email, password);
-      
-      // Verify admin access
-      const hasAdminAccess = await adminSecurity.verifyAdminAccess();
-      
-      if (!hasAdminAccess) {
-        await auth.signOut();
-        setError('Access denied. Admin privileges required.');
-        return;
-      }
-
-      // Generate admin session token
-      const sessionToken = await adminSecurity.generateAdminSessionToken();
-      
-      // Set admin session cookie (handled by server)
-      document.cookie = `admin-session=${sessionToken}; path=/; secure; samesite=strict; max-age=3600`;
-
-      toast({
-        title: 'Welcome back',
-        description: 'Admin access verified successfully',
-      });
-
-      // Redirect to intended page
-      router.push(redirect);
-    } catch (error: any) {
-      console.error('Admin login error:', error);
-      
-      if (error.code === 'auth/user-not-found' || error.code === 'auth/wrong-password') {
-        setError('Invalid email or password');
-      } else if (error.code === 'auth/too-many-requests') {
-        setError('Too many failed attempts. Please try again later.');
-      } else {
-        setError('Login failed. Please try again.');
-      }
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
   return (
-    <div className="min-h-screen flex items-center justify-center bg-gray-50 py-12 px-4 sm:px-6 lg:px-8">
-      <div className="max-w-md w-full space-y-8">
-        <div>
-          <div className="flex justify-center">
-            <Shield className="h-12 w-12 text-blue-600" />
-          </div>
-          <h2 className="mt-6 text-center text-3xl font-extrabold text-gray-900">
-            Admin Access
-          </h2>
-          <p className="mt-2 text-center text-sm text-gray-600">
-            Sign in to access the Dynasty admin dashboard
-          </p>
-        </div>
-
-        <form className="mt-8 space-y-6" onSubmit={handleSubmit}>
-          {error && (
-            <Alert variant="destructive">
-              <AlertCircle className="h-4 w-4" />
-              <AlertDescription>{error}</AlertDescription>
-            </Alert>
-          )}
-
-          <div className="space-y-4">
-            <div>
-              <Label htmlFor="email">Email address</Label>
-              <Input
-                id="email"
-                name="email"
-                type="email"
-                autoComplete="email"
-                required
-                value={email}
-                onChange={(e) => setEmail(e.target.value)}
-                className="mt-1"
-                placeholder="admin@mydynastyapp.com"
-              />
-            </div>
-
-            <div>
-              <Label htmlFor="password">Password</Label>
-              <Input
-                id="password"
-                name="password"
-                type="password"
-                autoComplete="current-password"
-                required
-                value={password}
-                onChange={(e) => setPassword(e.target.value)}
-                className="mt-1"
-              />
-            </div>
-          </div>
-
-          <div>
-            <Button
-              type="submit"
-              className="w-full"
-              disabled={isLoading}
-            >
-              {isLoading ? (
-                <>
-                  <Loader2 className="mr-2 h-4 w-4 animate-spin" />
-                  Signing in...
-                </>
-              ) : (
-                'Sign in'
-              )}
-            </Button>
-          </div>
-
-          <div className="text-center">
-            <Link
-              href="/"
-              className="text-sm text-blue-600 hover:text-blue-500"
-            >
-              Back to main site
-            </Link>
-          </div>
-        </form>
-
-        <div className="mt-8 p-4 bg-yellow-50 rounded-lg">
-          <p className="text-xs text-yellow-800">
-            <strong>Security Notice:</strong> This area is restricted to authorized administrators only. 
-            All access attempts are logged and monitored. Unauthorized access attempts will be reported.
-          </p>
-        </div>
-      </div>
-    </div>
+    <Suspense fallback={<div className="min-h-screen flex items-center justify-center"><div className="animate-spin rounded-full h-32 w-32 border-b-2 border-gray-900" /></div>}>
+      <AdminLoginClient />
+    </Suspense>
   );
 }
